requirements.txt
django>=1.8.3
pexpect
pyvmomi
djangorestframework
django-rest-swagger
django-angular
paramiko
ConfigObj
terminaltables